+++ src/mutex/mod.rs
@@ -6,26 +6,22 @@ mod async_mutex;
 #[cfg(feature = "alloc")]
 pub use async_mutex::*;
 
+mod atomic_mutex;
+pub use atomic_mutex::*;
+
 #[cfg(feature = "alloc")]
 mod mutex_alloc;
-
-#[cfg(feature = "std")]
-mod mutex_std;
-
-#[cfg(feature = "impl_parking_lot")]
-mod mutex_parking_lot;
-
-#[cfg(feature = "alloc")]
-mod park_mutex;
-
-mod spin_lock;
-
 #[cfg(feature = "alloc")]
 pub use mutex_alloc::*;
-
+#[cfg(feature = "impl_parking_lot")]
+mod mutex_parking_lot;
+#[cfg(feature = "std")]
+mod mutex_std;
+#[cfg(feature = "alloc")]
+mod park_mutex;
 #[cfg(feature = "alloc")]
 pub use park_mutex::*;
-
+mod spin_lock;
 pub use spin_lock::*;
 
 use crate::{EnsureSend, EnsureSync};